Search controller refactored, search service now redundant

+++ Services/BDInSelfLove.Services.Data/Comment/CommentService.cs
@@ -29,13 +29,13 @@
         public IQueryable<Comment> GetAllByArticleId(int articleId)
         {
             return this.commentRepository.All()
-                .Where(a => a.ArticleId == articleId);
+                .Where(c => c.ArticleId == articleId);
         }
 
         public IQueryable<Comment> GetAllByVideoId(int videoId)
         {
             return this.commentRepository.All()
-                .Where(vc => vc.VideoId == videoId);
+                .Where(c => c.VideoId == videoId);
         }
 
         public IQueryable<Comment> GetById(int commentId)
@@ -44,63 +44,55 @@
                 .Where(a => a.Id == commentId);
         }
 
-        public async Task<int> Edit(Comment serviceModel)
+        public async Task<int> Edit(Comment comment, string userId)
         {
-            var dbComment = await this.commentRepository.All().SingleOrDefaultAsync(a => a.Id == serviceModel.Id);
+            var dbComment = await this.commentRepository.All().Where(c => c.Id == comment.Id).SingleOrDefaultAsync();
 
-            if (dbComment == null)
+            // Check if comment exists & creator is same as editor
+            if (dbComment == null || dbComment.UserId != userId)
             {
                 return 0;
             }
 
-            dbComment.Content = serviceModel.Content;
-
+            dbComment.Content = comment.Content;
             this.commentRepository.Update(dbComment);
-            int result = await this.commentRepository.SaveChangesAsync();
 
-            return result;
+            return await this.commentRepository.SaveChangesAsync();
         }
 
-        public async Task<int> Delete(int commentId)
+        public async Task<int> Delete(int commentId, string userId, bool isUserAdmin)
         {
+            var comment = await this.commentRepository.All().SingleOrDefaultAsync(c => c.Id == commentId);
+
+            if (comment == null || (comment.UserId != userId && !isUserAdmin))
+            {
+                // Comment doesn't exist || user not same as creator and not admin
+                return 0;
+            }
+
             var secondLevelComments = await this.commentRepository.All().Where(c => c.ParentCommentId == commentId).ToListAsync();
 
-            foreach (var comment in secondLevelComments)
+            foreach (var nestedComment in secondLevelComments)
             {
                 // Delete most deeply nested comments
-                var thirdLevelComments = await this.commentRepository.All().Where(c => c.ParentCommentId == comment.Id).ToListAsync();
+                var thirdLevelComments = await this.commentRepository.All().Where(c => c.ParentCommentId == nestedComment.Id).ToListAsync();
                 foreach (var lastLevelComment in thirdLevelComments)
                 {
                     this.commentRepository.Delete(lastLevelComment);
                 }
 
                 // Delete second level comments
-                this.commentRepository.Delete(comment);
+                this.commentRepository.Delete(nestedComment);
             }
 
-            // Delete selected comment
-            var selectedComment = await this.commentRepository.All().SingleOrDefaultAsync(c => c.Id == commentId);
-
-            // Check whether invalid comment id was entered manually by user
-            if (selectedComment == null)
-            {
-                return 0;
-            }
-
-            this.commentRepository.Delete(selectedComment);
+            // Delete comment
+            this.commentRepository.Delete(comment);
             int result = await this.commentRepository.SaveChangesAsync();
-
             return result;
         }
 
-        public ICollection<CommentServiceModel> ArrangeCommentHierarchy(ICollection<CommentServiceModel> comments)
+        public ICollection<Comment> ArrangeCommentHierarchy(ICollection<Comment> comments)
         {
-            //// Clear ef subcomment structure as it only goes 1 level deep
-            //foreach (var comment in comments)
-            //{
-            //    comment.SubComments.Clear();
-            //}
-
             // Populate subcomments references
             foreach (var comment in comments.Where(c => c.ParentCommentId != null))
             {
